Create package for serialization operations
