+++ src/main/java/org/duraspace/bagit/serialize/BagDeserializer.java
@@ -2,7 +2,7 @@
  * The contents of this file are subject to the license and copyright detailed
  * in the LICENSE and NOTICE files at the root of the source tree.
  */
-package org.duraspace.bagit;
+package org.duraspace.bagit.serialize;
 
 import java.io.IOException;
 import java.nio.file.Path;
