+++ src/main/java/org/duraspace/bagit/serialize/GZipBagDeserializer.java
@@ -2,7 +2,7 @@
  * The contents of this file are subject to the license and copyright detailed
  * in the LICENSE and NOTICE files at the root of the source tree.
  */
-package org.duraspace.bagit;
+package org.duraspace.bagit.serialize;
 
 import java.io.BufferedInputStream;
 import java.io.IOException;
@@ -12,6 +12,7 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 
 import org.apache.commons.compress.compressors.gzip.GzipCompressorInputStream;
+import org.duraspace.bagit.BagProfile;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
