+++ src/main/java/org/duraspace/bagit/serialize/SerializationSupport.java
@@ -2,7 +2,7 @@
  * The contents of this file are subject to the license and copyright detailed
  * in the LICENSE and NOTICE files at the root of the source tree.
  */
-package org.duraspace.bagit;
+package org.duraspace.bagit.serialize;
 
 import java.io.IOException;
 import java.io.UncheckedIOException;
@@ -16,6 +16,7 @@ import java.util.Set;
 
 import org.apache.tika.Tika;
 import org.duraspace.bagit.exception.BagProfileException;
+import org.duraspace.bagit.BagProfile;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
