+++ src/test/java/org/duraspace/bagit/serialize/BagSerializerTest.java
@@ -2,7 +2,7 @@
  * The contents of this file are subject to the license and copyright detailed
  * in the LICENSE and NOTICE files at the root of the source tree.
  */
-package org.duraspace.bagit;
+package org.duraspace.bagit.serialize;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -20,6 +20,9 @@ import java.util.zip.GZIPInputStream;
 import org.apache.commons.compress.archivers.ArchiveEntry;
 import org.apache.commons.compress.archivers.tar.TarArchiveInputStream;
 import org.apache.commons.compress.archivers.zip.ZipArchiveInputStream;
+import org.duraspace.bagit.BagProfile;
+import org.duraspace.bagit.serialize.BagSerializer;
+import org.duraspace.bagit.serialize.SerializationSupport;
 import org.junit.Before;
 import org.junit.Test;
